barcode-service/requirements.txt
fastapi
uvicorn
python-multipart
pyzbar
Pillow
httpx
